+++ src/theme/dark.ts
@@ -352,6 +352,30 @@ const chatMessage = {
   }
 }
 
+const channelList = {
+  bgd: colors.greys.grey80,
+  fontColor: colors.greys.grey10,
+  border: '1px solid transparent',
+  active: {
+    bgd: colors.primary.dark,
+    fontColor: colors.greys.grey70,
+  },
+  hover: {
+    bgd: colors.greys.grey70,
+  },
+  focus: {
+    border: `1px solid ${colors.primary.dark}`,
+    selectedBorder: `1px solid ${colors.greys.grey70}`,
+  },
+  selected: {
+    bgd: colors.primary.light,
+    fontColor: colors.greys.grey70,
+  },
+  iconButton: {
+    activeBgd: colors.greys.grey80,
+  }
+}
+
 export const darkTheme = {
   name: 'Dark Theme',
   buttons,
@@ -368,6 +392,7 @@ export const darkTheme = {
   navbar,
   videoGrid,
   chatMessage,
+  channelList,
   ...defaultTheme
 };