+++ src/theme/light.ts
@@ -350,7 +350,31 @@ const chatMessage = {
     fontColor: colors.greys.grey70,
     bgd: colors.greys.grey20,
   }
-};
+}
+
+const channelList = {
+  bgd: colors.greys.white,
+  fontColor: colors.greys.grey70,
+  border: '1px solid transparent',
+  active: {
+    bgd: colors.primary.dark,
+    fontColor: colors.greys.white,
+  },
+  hover: {
+    bgd: colors.greys.grey10,
+  },
+  focus: {
+    border: `1px solid ${colors.primary.dark}`,
+    selectedBorder: `1px solid ${colors.greys.grey10}`,
+  },
+  selected: {
+    bgd: colors.primary.light,
+    fontColor: colors.greys.white,
+  },
+  iconButton: {
+    activeBgd: colors.greys.white,
+  }
+}
 
 export const lightTheme = {
   name: 'Light Theme',
@@ -368,6 +392,7 @@ export const lightTheme = {
   navbar,
   videoGrid,
   chatMessage,
+  channelList,
   ...defaultTheme
 };