+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add chat message component
 - Add `isSelected` prop to `ControlBarButton` component
 - Add `UpAndDownCaret` icon component
+- Add channel list component for chat
 
 ### Changed